feat/refactor: extract shared utils & add _get_task_dependency_parent for dag+taskgroup
