+++ orbiter/objects/task.py
@@ -2,12 +2,10 @@ from __future__ import annotations
 
 import ast
 from abc import ABC
-from typing import Set, List, Annotated, Callable, Literal, TYPE_CHECKING
+from typing import Set, List, Callable, Literal
 
-from loguru import logger
-from pydantic import AfterValidator, BaseModel, validate_call
+from pydantic import BaseModel
 
-from orbiter import clean_value
 from orbiter.ast_helper import (
     OrbiterASTBase,
     py_function,
@@ -19,9 +17,7 @@ from orbiter.objects import ImportList, RenderAttributes, CALLBACK_KEYS
 from orbiter.objects import OrbiterBase
 from orbiter.objects.callbacks.callback_type import CallbackType
 from orbiter.objects.pool import OrbiterPool
-
-if TYPE_CHECKING:
-    from orbiter.objects.task_group import OrbiterTaskGroup
+from orbiter.objects.task_shared_utils import TaskId, task_add_downstream, to_task_id
 
 __mermaid__ = """
 --8<-- [start:mermaid-dag-relationships]
@@ -37,57 +33,6 @@ OrbiterOperator "implements" <|-- OrbiterTask
 OrbiterOperator --> "many" OrbiterCallback
 --8<-- [end:mermaid-task-relationships]
 """
-
-TaskId = Annotated[str, AfterValidator(lambda t: to_task_id(t))]
-
-
-def task_add_downstream(self, task_id: str | List[str] | OrbiterTaskDependency) -> "OrbiterOperator | OrbiterTaskGroup":
-    # noinspection PyProtectedMember
-    """
-    Add a downstream task dependency
-
-    ```pycon
-    >>> from orbiter.objects.operators.empty import OrbiterEmptyOperator
-    >>> from orbiter.objects.task import task_add_downstream, OrbiterTaskDependency
-    >>> from orbiter.ast_helper import render_ast
-    >>> render_ast(task_add_downstream(OrbiterEmptyOperator(task_id="task_id"), "downstream")._downstream_to_ast())
-    'task_id_task >> downstream_task'
-    >>> render_ast(
-    ...     task_add_downstream(OrbiterEmptyOperator(task_id="task_id"),
-    ...     OrbiterTaskDependency(task_id="task_id", downstream="downstream"))._downstream_to_ast()
-    ... )
-    'task_id_task >> downstream_task'
-    >>> render_ast(task_add_downstream(
-    ...     OrbiterEmptyOperator(task_id="task_id"),
-    ...     ["downstream"]
-    ... )._downstream_to_ast())
-    'task_id_task >> downstream_task'
-    >>> render_ast(task_add_downstream(
-    ...     OrbiterEmptyOperator(task_id="task_id"),
-    ...     ["downstream", "downstream2"]
-    ... )._downstream_to_ast())
-    'task_id_task >> [downstream2_task, downstream_task]'
-
-    ```
-    """  # noqa: E501
-    if isinstance(task_id, OrbiterTaskDependency):
-        task_dependency = task_id
-        if task_dependency.task_id != self.task_id:
-            raise ValueError(f"task_dependency={task_dependency} has a different task_id than {self.task_id}")
-        # do normal parsing logic, but with these downstream items
-        task_id = task_dependency.downstream
-
-    if isinstance(task_id, str):
-        self.downstream |= {to_task_id(task_id)}
-        return self
-    else:
-        if not len(task_id):
-            return self
-
-        downstream_task_id = {to_task_id(t) for t in task_id}
-        logger.debug(f"Adding downstream {downstream_task_id} to {self.task_id}")
-        self.downstream |= downstream_task_id
-        return self
 
 
 class OrbiterTaskDependency(BaseModel, extra="forbid"):
@@ -340,25 +285,3 @@ class OrbiterTask(OrbiterOperator, extra="allow"):
             if len(callable_props)
             else self_as_ast
         )
-
-
-@validate_call
-def to_task_id(task_id: str, assignment_suffix: str = "") -> str:
-    # noinspection PyTypeChecker
-    """General utiltty function - turns MyTaskId into my_task_id (or my_task_id_task suffix is `_task`)
-    ```pycon
-    >>> to_task_id("MyTaskId")
-    'my_task_id'
-    >>> to_task_id("MyTaskId", "_task")
-    'my_task_id_task'
-    >>> to_task_id("my_task_id_task", "_task")
-    'my_task_id_task'
-
-    ```
-    :param task_id:
-    :type task_id: str
-    :param assignment_suffix: e.g. `_task` for `task_id_task = MyOperator(...)`
-    :type assignment_suffix: str
-    """
-    task_id = clean_value(task_id)
-    return task_id + (assignment_suffix if task_id[-len(assignment_suffix) :] != assignment_suffix else "")
